feat: convert response to content fetching
With custom error enum `RusTurengError`, convert the response fetching
process, to a content fetching process. To handle the possible
access errors in one place, any `reqwest::Error`, not `200 Ok` response
from the request creates a corresponding `RusTurengError`. With this
framework, content fetching is either returns a access related error, or
the content itself. From this point, problem reduces to parsing the
provided content.