feat: add basic translation result
The translation request can result in three different results. The first
one is the `Term Not Found` result, which corresponds to the case where
the searched word is too irrelevant, or incorrect that result is just
not found. The second one is the suggestions, the searched term is close
to another valid word, and off by maybe a typo, or so. In this case,
list of suggestions are returned. Finally, for the valid search case,
two tables are returned with valid translations.