feat: convert builder `Request` to `HttpClient`
With the `Request` builder the program was generating `possibly lost`
memory leaks due to global allocation strategy of `isahc` crate (see
issue sagebind/isahc#427). Instead of this approach, an `HttpClient` is
generated and requests are handled with this client. With the
`HttpClient` approach, output reduces to `still reachable` leak, which
is expected.