fix(parser): add work-around for new `sentences`
The `sentences` and `history` option is added to the UI of `tureng.com`
which adds additional sentence examples for the translations. They are
shipped as separate table entries which do not start with an index. This
new addition was breaking the parser which relies on the existence of
the indices.

With this fix, the sentence information is filtered out by only
returning the table entries starting with an index. This behavior
eliminates the `sentences` information for now. A future implementation
would be necessary to employ additional information.